put the buggy version instead of the correct version in the bench

+++ benchmark/bugsjar/camel/9319e139/buggy-version/camel-core/src/main/java/org/apache/camel/component/file/strategy/MarkerFileExclusiveReadLockStrategy.java
@@ -34,6 +34,8 @@ import org.slf4j.LoggerFactory;
  */
 public class MarkerFileExclusiveReadLockStrategy implements GenericFileExclusiveReadLockStrategy<File> {
     private static final transient Logger LOG = LoggerFactory.getLogger(MarkerFileExclusiveReadLockStrategy.class);
+    private File lock;
+    private String lockFileName;
 
     public void prepareOnStartup(GenericFileOperations<File> operations, GenericFileEndpoint<File> endpoint) {
         String dir = endpoint.getConfiguration().getDirectory();
@@ -48,25 +50,28 @@ public class MarkerFileExclusiveReadLockStrategy implements GenericFileExclusive
 
     public boolean acquireExclusiveReadLock(GenericFileOperations<File> operations,
                                             GenericFile<File> file, Exchange exchange) throws Exception {
-        String lockFileName = getLockFileName(file);
+        lockFileName = file.getAbsoluteFilePath() + FileComponent.DEFAULT_LOCK_FILE_POSTFIX;
         LOG.trace("Locking the file: {} using the lock file name: {}", file, lockFileName);
 
         // create a plain file as marker filer for locking (do not use FileLock)
-        File lock = new File(lockFileName);
+        lock = new File(lockFileName);
         boolean acquired = lock.createNewFile();
+        if (!acquired) {
+            lock = null;
+
+        }
 
         return acquired;
     }
 
     public void releaseExclusiveReadLock(GenericFileOperations<File> operations,
                                          GenericFile<File> file, Exchange exchange) throws Exception {
-        String lockFileName = getLockFileName(file);
-        File lock = new File(lockFileName);
+        if (lock != null) {
+            LOG.trace("Unlocking file: {}", lockFileName);
 
-        LOG.trace("Unlocking file: {}", lockFileName);
-
-        boolean deleted = FileUtil.deleteFile(lock);
-        LOG.trace("Lock file: {} was deleted: {}", lockFileName, deleted);
+            boolean deleted = FileUtil.deleteFile(lock);
+            LOG.trace("Lock file: {} was deleted: {}", lockFileName, deleted);
+        }
     }
 
     public void setTimeout(long timeout) {
@@ -94,10 +99,6 @@ public class MarkerFileExclusiveReadLockStrategy implements GenericFileExclusive
                 deleteLockFiles(file, true);
             }
         }
-    }
-
-    private static String getLockFileName(GenericFile<File> file) {
-        return file.getAbsoluteFilePath() + FileComponent.DEFAULT_LOCK_FILE_POSTFIX;
     }
 
 }
